backend/pyproject.toml
[project]
name = "gemini-chat-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "google-generativeai>=0.7.0",
    "mem0ai>=0.0.20",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.0.0",
]